Move get_execute_command to shell.py, rename get_interpreter_for_command to get_interpreter, and add tests it.
