+++ src/python/bot/startup/run.py
@@ -177,13 +177,12 @@ def main():
   bot_log = os.path.join(log_directory, 'bot.log')
 
   bot_script_path = os.path.join(base_directory, BOT_SCRIPT)
-  bot_interpreter = shell.get_interpreter_for_command(bot_script_path)
+  bot_interpreter = shell.get_interpreter(bot_script_path)
   assert bot_interpreter
   bot_command = '%s %s' % (bot_interpreter, bot_script_path)
 
   heartbeat_script_path = os.path.join(base_directory, HEARTBEAT_SCRIPT)
-  heartbeat_interpreter = shell.get_interpreter_for_command(
-      heartbeat_script_path)
+  heartbeat_interpreter = shell.get_interpreter(heartbeat_script_path)
   assert heartbeat_interpreter
   heartbeat_command = '%s %s %s' % (heartbeat_interpreter,
                                     heartbeat_script_path, bot_log)
