+++ src/python/bot/tasks/setup.py
@@ -471,7 +471,7 @@ def update_fuzzer_and_data_bundles(fuzzer_name):
 
   # Setup environment variable for launcher script path.
   if fuzzer.launcher_script:
-    fuzzer_launcher_path = utils.get_execute_command(
+    fuzzer_launcher_path = shell.get_execute_command(
         os.path.join(fuzzer_directory, fuzzer.launcher_script))
     environment.set_value('LAUNCHER_PATH', fuzzer_launcher_path)
